requirements.txt
tracr-pypi
torch
plotly
matplotlib
transformer-lens
numpy